Cisco Content Services Switch - from "Wedge Martin" <wedge@shopping.com>
"I call this 'cssrancid' and use the vendor type of 'css' to
differentiate it.
It's a total hack on the standard rancid script, but it has all the
functional differences, including one that was super hard to track
down...

The user profile, when the term length is changed to 65535 ( css handles
term len 0 stupidly and gives you a term len of 24 ) it prompts you when
you log out to commit or discard user profile changes, screwing up the
session by hanging indefinitely...  so the trick is to copy the profile
to user-profile; i found this _buried_ in cisco documentation..   goofy
voodoo.. but it works.  i threw in a couple of other little hacks to
make this functional, and it's working nicely..

feel free to put this in the next distribution...  i'll clean it up a
bit too and send you any changes i make.
"