clogin.in: catch More prompts for IOS-XR in admin mode
When using "admin show ...", the "more" prompt is in reverse video. It
is not possible to disable it with "terminal length 0" when running as
a read-only user (dunno why exactly, it works with a fully privileged
user).

- `^[[7m--More--^[[27m`
- `^[[7m(END)^[[27m`

Catch them both. We do not require to start at the beginning because
"(END)" prompt can replace the "--More--" prompt without displaying
any new line.